+++ sys/src/vista/formularios/FrmSysGrupo.java
@@ -3,6 +3,9 @@ package vista.formularios;
 import java.util.ArrayList;
 import java.util.List;
 
+import javax.swing.DefaultCellEditor;
+import javax.swing.JCheckBox;
+import javax.swing.JComboBox;
 import javax.swing.JScrollPane;
 import javax.swing.JTable;
 import javax.swing.ListSelectionModel;
@@ -10,12 +13,15 @@ import javax.swing.ListSelectionModel;
 import vista.contenedores.ctnModulo;
 import vista.contenedores.ctnTitulo;
 import vista.controles.DSGTableModel;
+import vista.utilitarios.ButtonRenderer;
 import vista.utilitarios.MaestroTableModel;
+import vista.utilitarios.UtilMensajes;
 
 import javax.swing.JLabel;
 import javax.swing.JTextField;
 import javax.swing.event.ListSelectionEvent;
 import javax.swing.event.ListSelectionListener;
+import javax.swing.table.TableCellEditor;
 import javax.swing.JButton;
 
 import dao.SysGrupoDAO;
@@ -31,6 +37,8 @@ import entity.SysTitulo;
 
 import java.awt.event.ActionListener;
 import java.awt.event.ActionEvent;
+import java.awt.event.FocusAdapter;
+import java.awt.event.FocusEvent;
 
 import javax.swing.GroupLayout;
 import javax.swing.GroupLayout.Alignment;
@@ -51,8 +59,13 @@ public class FrmSysGrupo extends AbstractMaestro {
 	private JTable tblOpciones;
 	private JTextField txtCodigo;
 	private JTextField txtDescripcion;
+	
+	private JComboBox combo1;
 
 	private SysGrupo sysGrupo;
+	private SysGrupo sysGrupo2;
+	private SysGrupoPK sysGrupoPK;
+	private SysTitulo sysTitulo;
 	private SysModulo sysModulo;
 	private SysModuloDAO sysModuloDAO = new SysModuloDAO();
 	private SysTituloDAO sysTituloDAO = new SysTituloDAO();
@@ -61,7 +74,6 @@ public class FrmSysGrupo extends AbstractMaestro {
 	private List<SysOpcion> sysOpciones;
 
 	private SysGrupoDAO sysGrupoDAO = new SysGrupoDAO();
-
 	private SysOpcionDAO sysOpcionDAO = new SysOpcionDAO();
 
 	private JButton btnILinea;
@@ -74,6 +86,11 @@ public class FrmSysGrupo extends AbstractMaestro {
 
 		JScrollPane scrollPane = new JScrollPane();
 
+		combo1=new JComboBox();
+		combo1.addItem("grande");
+        combo1.addItem("mediano");
+        combo1.addItem("pequeño");	
+        
 		tblLista = new JTable(new MaestroTableModel());
 		scrollPane.setViewportView(tblLista);
 		tblLista.setSelectionMode(ListSelectionModel.SINGLE_SELECTION);
@@ -92,10 +109,15 @@ public class FrmSysGrupo extends AbstractMaestro {
 			public boolean evaluaEdicion(int row, int column) {
 				return getEditar();
 			}
+			
+			
 		});
 		scrollPaneNum.setViewportView(tblOpciones);
 		tblOpciones.setSelectionMode(ListSelectionModel.SINGLE_SELECTION);
-
+			
+		DefaultCellEditor defaultCellEditor=new DefaultCellEditor(combo1);
+		tblOpciones.getColumn("Prioridad").setCellEditor(defaultCellEditor);
+		
 		JLabel lblCdigo = new JLabel("C\u00F3d Grupo");
 
 		JLabel lblDescripcin = new JLabel("Descripci\u00F3n");
@@ -109,7 +131,8 @@ public class FrmSysGrupo extends AbstractMaestro {
 		btnILinea = new JButton("I Linea");
 		btnILinea.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent arg0) {
-				final Object fila[] = { "", "", "", "", "" };
+				
+				final Object fila[] = { idMas(), "", "", "", };
 				getDetalleTM().addRow(fila);
 			}
 		});
@@ -123,13 +146,22 @@ public class FrmSysGrupo extends AbstractMaestro {
 			}
 		});
 		
-		ctnmodulo = new ctnModulo();
+		//ctnmodulo = new ctnModulo(new SysModuloDAO());
 		//txtidmodulo.setColumns(10);
 		
 		lblCdModulo = new JLabel("C\u00F3d. Modulo");
 		
+		ctnmodulo = new ctnModulo(new SysModuloDAO());
+		
 		ctntitulo = new ctnTitulo();
+		ctntitulo.setSysTituloDAO(new SysTituloDAO());
 		//txtidtitulo.setColumns(10);
+		ctntitulo.txtCodigo.addFocusListener(new FocusAdapter() {
+			@Override
+			public void focusGained(FocusEvent arg0) {
+				actualizaTitulo();
+			}
+		});
 		
 		lblCdigo_1 = new JLabel("C\u00F3d Titulo");
 		GroupLayout groupLayout = new GroupLayout(pnlContenido);
@@ -137,38 +169,46 @@ public class FrmSysGrupo extends AbstractMaestro {
 			groupLayout.createParallelGroup(Alignment.LEADING)
 				.addGroup(groupLayout.createSequentialGroup()
 					.addContainerGap()
-					.addComponent(scrollPane, GroupLayout.PREFERRED_SIZE, 199, GroupLayout.PREFERRED_SIZE)
+					.addComponent(scrollPane, GroupLayout.DEFAULT_SIZE, 199, Short.MAX_VALUE)
 					.addGap(10)
 					.addGroup(groupLayout.createParallelGroup(Alignment.LEADING)
-						.addComponent(scrollPaneNum, GroupLayout.PREFERRED_SIZE, 314, GroupLayout.PREFERRED_SIZE)
 						.addGroup(groupLayout.createSequentialGroup()
-							.addGroup(groupLayout.createParallelGroup(Alignment.LEADING, false)
-								.addGroup(groupLayout.createSequentialGroup()
-									.addComponent(lblCdModulo, GroupLayout.PREFERRED_SIZE, 66, GroupLayout.PREFERRED_SIZE)
-									.addGap(13)
-									.addComponent(ctnmodulo, GroupLayout.DEFAULT_SIZE, GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-								.addGroup(groupLayout.createSequentialGroup()
-									.addComponent(lblCdigo_1, GroupLayout.PREFERRED_SIZE, 66, GroupLayout.PREFERRED_SIZE)
-									.addGap(13)
-									.addComponent(ctntitulo, GroupLayout.DEFAULT_SIZE, GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-								.addGroup(groupLayout.createSequentialGroup()
-									.addComponent(lblDescripcin, GroupLayout.PREFERRED_SIZE, 66, GroupLayout.PREFERRED_SIZE)
-									.addGap(13)
-									.addComponent(txtDescripcion, GroupLayout.PREFERRED_SIZE, 126, GroupLayout.PREFERRED_SIZE))
+							.addComponent(scrollPaneNum, GroupLayout.DEFAULT_SIZE, 314, Short.MAX_VALUE)
+							.addGap(6))
+						.addGroup(groupLayout.createSequentialGroup()
+							.addGroup(groupLayout.createParallelGroup(Alignment.LEADING)
 								.addGroup(groupLayout.createSequentialGroup()
 									.addComponent(lblCdigo, GroupLayout.PREFERRED_SIZE, 66, GroupLayout.PREFERRED_SIZE)
 									.addGap(13)
-									.addComponent(txtCodigo, GroupLayout.PREFERRED_SIZE, 67, GroupLayout.PREFERRED_SIZE)))
-							.addPreferredGap(ComponentPlacement.RELATED, 26, Short.MAX_VALUE)
-							.addGroup(groupLayout.createParallelGroup(Alignment.LEADING, false)
-								.addComponent(btnILinea, GroupLayout.DEFAULT_SIZE, GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-								.addComponent(btnBLinea, GroupLayout.DEFAULT_SIZE, 83, Short.MAX_VALUE))))
-					.addContainerGap())
+									.addComponent(txtCodigo, GroupLayout.PREFERRED_SIZE, 67, GroupLayout.PREFERRED_SIZE)
+									.addGap(85))
+								.addGroup(groupLayout.createSequentialGroup()
+									.addGroup(groupLayout.createParallelGroup(Alignment.TRAILING)
+										.addGroup(groupLayout.createSequentialGroup()
+											.addComponent(lblCdigo_1, GroupLayout.PREFERRED_SIZE, 66, GroupLayout.PREFERRED_SIZE)
+											.addGap(13)
+											.addComponent(ctntitulo, GroupLayout.DEFAULT_SIZE, GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+										.addGroup(Alignment.LEADING, groupLayout.createSequentialGroup()
+											.addComponent(lblCdModulo, GroupLayout.PREFERRED_SIZE, 66, GroupLayout.PREFERRED_SIZE)
+											.addGap(13)
+											.addComponent(ctnmodulo, GroupLayout.PREFERRED_SIZE, 126, Short.MAX_VALUE))
+										.addGroup(Alignment.LEADING, groupLayout.createSequentialGroup()
+											.addComponent(lblDescripcin, GroupLayout.PREFERRED_SIZE, 66, GroupLayout.PREFERRED_SIZE)
+											.addGap(13)
+											.addComponent(txtDescripcion, GroupLayout.PREFERRED_SIZE, 144, GroupLayout.PREFERRED_SIZE)))
+									.addGap(8)))
+							.addGroup(groupLayout.createParallelGroup(Alignment.LEADING)
+								.addComponent(btnBLinea, GroupLayout.PREFERRED_SIZE, 83, GroupLayout.PREFERRED_SIZE)
+								.addComponent(btnILinea, GroupLayout.PREFERRED_SIZE, 83, GroupLayout.PREFERRED_SIZE))
+							.addGap(6))))
 		);
 		groupLayout.setVerticalGroup(
 			groupLayout.createParallelGroup(Alignment.LEADING)
-				.addGroup(groupLayout.createSequentialGroup()
-					.addGroup(groupLayout.createParallelGroup(Alignment.LEADING)
+				.addGroup(Alignment.TRAILING, groupLayout.createSequentialGroup()
+					.addGroup(groupLayout.createParallelGroup(Alignment.TRAILING)
+						.addGroup(Alignment.LEADING, groupLayout.createSequentialGroup()
+							.addContainerGap()
+							.addComponent(scrollPane, GroupLayout.DEFAULT_SIZE, 239, Short.MAX_VALUE))
 						.addGroup(groupLayout.createSequentialGroup()
 							.addGroup(groupLayout.createParallelGroup(Alignment.LEADING)
 								.addGroup(groupLayout.createSequentialGroup()
@@ -176,22 +216,24 @@ public class FrmSysGrupo extends AbstractMaestro {
 									.addComponent(lblCdModulo))
 								.addGroup(groupLayout.createSequentialGroup()
 									.addGap(9)
-									.addComponent(ctnmodulo, GroupLayout.PREFERRED_SIZE, GroupLayout.DEFAULT_SIZE, GroupLayout.PREFERRED_SIZE)))
+									.addComponent(ctnmodulo, GroupLayout.PREFERRED_SIZE, 18, GroupLayout.PREFERRED_SIZE)))
+							.addGap(6)
 							.addGroup(groupLayout.createParallelGroup(Alignment.LEADING)
 								.addGroup(groupLayout.createSequentialGroup()
 									.addGap(8)
 									.addComponent(lblCdigo_1))
 								.addGroup(groupLayout.createSequentialGroup()
 									.addPreferredGap(ComponentPlacement.RELATED)
-									.addComponent(ctntitulo, GroupLayout.PREFERRED_SIZE, GroupLayout.DEFAULT_SIZE, GroupLayout.PREFERRED_SIZE)))
+									.addComponent(ctntitulo, GroupLayout.PREFERRED_SIZE, 18, GroupLayout.PREFERRED_SIZE)))
 							.addPreferredGap(ComponentPlacement.RELATED)
 							.addGroup(groupLayout.createParallelGroup(Alignment.LEADING)
 								.addComponent(lblCdigo)
 								.addGroup(groupLayout.createSequentialGroup()
 									.addGap(4)
-									.addGroup(groupLayout.createParallelGroup(Alignment.BASELINE)
-										.addComponent(txtCodigo, GroupLayout.PREFERRED_SIZE, GroupLayout.DEFAULT_SIZE, GroupLayout.PREFERRED_SIZE)
-										.addComponent(btnILinea, GroupLayout.PREFERRED_SIZE, 20, GroupLayout.PREFERRED_SIZE))))
+									.addComponent(txtCodigo, GroupLayout.PREFERRED_SIZE, GroupLayout.DEFAULT_SIZE, GroupLayout.PREFERRED_SIZE))
+								.addGroup(groupLayout.createSequentialGroup()
+									.addGap(4)
+									.addComponent(btnILinea, GroupLayout.PREFERRED_SIZE, 20, GroupLayout.PREFERRED_SIZE)))
 							.addGap(4)
 							.addGroup(groupLayout.createParallelGroup(Alignment.LEADING)
 								.addGroup(groupLayout.createSequentialGroup()
@@ -201,11 +243,8 @@ public class FrmSysGrupo extends AbstractMaestro {
 									.addComponent(txtDescripcion, GroupLayout.PREFERRED_SIZE, GroupLayout.DEFAULT_SIZE, GroupLayout.PREFERRED_SIZE)
 									.addComponent(btnBLinea, GroupLayout.PREFERRED_SIZE, 20, GroupLayout.PREFERRED_SIZE)))
 							.addGap(11)
-							.addComponent(scrollPaneNum, GroupLayout.DEFAULT_SIZE, 130, Short.MAX_VALUE))
-						.addGroup(groupLayout.createSequentialGroup()
-							.addContainerGap()
-							.addComponent(scrollPane, GroupLayout.DEFAULT_SIZE, 235, Short.MAX_VALUE)))
-					.addContainerGap())
+							.addComponent(scrollPaneNum, GroupLayout.DEFAULT_SIZE, 124, Short.MAX_VALUE)))
+					.addGap(7))
 		);
 		pnlContenido.setLayout(groupLayout);
 		tblLista.getSelectionModel().addListSelectionListener(
@@ -213,14 +252,36 @@ public class FrmSysGrupo extends AbstractMaestro {
 					@Override
 					public void valueChanged(ListSelectionEvent e) {
 						int selectedRow = tblLista.getSelectedRow();
-						if (selectedRow >= 0)
+						if (selectedRow >= 0){
 							setSysGrupo(getSysGrupos().get(selectedRow));
+							setSysGrupo2(getSysGrupos().get(selectedRow));
+							setSysTitulo(getSysGrupos().get(selectedRow).getSysTitulo());
+						}							
 						else
 							setSysGrupo(null);
 						llenar_datos();
 					}
 				});
 		iniciar();
+	}
+	
+	public void actualizaTitulo(){
+		ctntitulo.setModulo(ctnmodulo.getSeleccionado());
+		ctntitulo.refrescar();
+	}
+	
+	public String idMas(){
+		int tot;
+		
+		tot = tblOpciones.getRowCount() + 1;
+		
+		if(tot<10){
+			return "00"+tot;
+		}else if(tot < 100){
+			return "0"+tot;
+		}else{
+			return ""+tot;
+		}
 	}
 
 	@Override
@@ -251,11 +312,13 @@ public class FrmSysGrupo extends AbstractMaestro {
 		for (SysOpcion obj : getSysOpciones()) {
 			getSysOpcionDAO().create(obj);
 		}
+		//iniciar();
 	}
 
 	@Override
 	public void llenarDesdeVista() {
-		String idmodulo, idtitulo, idgrupo;
+		String idmodulo, idtitulo, idgrupo;		
+		
 		idmodulo = ctnmodulo.getSeleccionado().getIdmodulo();
 		idtitulo = ctntitulo.getSeleccionado().getId().getIdtitulo();
 		idgrupo = this.txtCodigo.getText().trim();
@@ -280,8 +343,15 @@ public class FrmSysGrupo extends AbstractMaestro {
 
 			obj.setId(ido);
 			obj.setDescripcion(getDetalleTM().getValueAt(i, 1).toString());
-			System.out.println(getDetalleTM().getValueAt(i, 2));
-			obj.setPrioridad(Integer.parseInt(getDetalleTM().getValueAt(i, 2).toString()));
+			//System.out.println(getDetalleTM().getValueAt(i, 2));
+			if(getDetalleTM().getValueAt(i, 2).equals("grande")){
+				obj.setPrioridad(1);
+			}else if(getDetalleTM().getValueAt(i, 2).equals("mediano")){
+				obj.setPrioridad(2);
+			}else if(getDetalleTM().getValueAt(i, 2).equals("pequeño")){
+				obj.setPrioridad(3);
+			}
+			
 			obj.setImagen(getDetalleTM().getValueAt(i, 3).toString());
 			obj.setOpcion(getDetalleTM().getValueAt(i, 4).toString());
 			getSysOpciones().add(obj);
@@ -291,28 +361,34 @@ public class FrmSysGrupo extends AbstractMaestro {
 
 	@Override
 	public void llenar_datos() {
+		
 		getDetalleTM().limpiar();
-		setSysOpciones(new ArrayList<SysOpcion>());
-		if (getSysGrupo() != null) {
+		//setSysOpciones(new ArrayList<SysOpcion>());
+		
+		if (!getEstado().equals(NUEVO) && sysGrupoDAO.findAll().size()>0) {
 			
-			for(SysModulo sm : sysModuloDAO.findAll()){
-				if(sm.getIdmodulo().equals(getSysGrupo().getId().getIdmodulo())){
-					ctnmodulo.setSeleccionado(sm);
-				}
-			}			
-			for(SysTitulo st : sysTituloDAO.findAll()){
-				if(st.getId().getIdtitulo().equals(getSysGrupo().getId().getIdtitulo())){
-					ctntitulo.setSeleccionado(st);
-				}
-			}			
-			txtCodigo.setText(getSysGrupo().getId().getIdgrupo());
-			txtDescripcion.setText(getSysGrupo().getDescripcion());
-			setSysOpciones(getSysOpcionDAO().getPorGrupo(getSysGrupo()));
+			ctnmodulo.setSeleccionado(getSysTitulo().getSysModulo());		
+			ctntitulo.setSeleccionado(getSysTitulo());
+			
+			System.out.println("Mostrar: "+getSysGrupo2().getId().getIdgrupo());
+			txtCodigo.setText(getSysGrupo2().getId().getIdgrupo().toString());
+			txtDescripcion.setText(getSysGrupo2().getDescripcion());
+			setSysOpciones(getSysOpcionDAO().getPorGrupo(getSysGrupo2()));
 
 			for (SysOpcion obj : getSysOpciones()) {
+				
+				String op;
+				if(obj.getPrioridad() == 1){
+					op = "grande";
+				}else if(obj.getPrioridad() == 2){
+					op = "mediano";
+				}else{
+					op = "pequeño";
+				}
+				
 				getDetalleTM().addRow(
 						new Object[] { obj.getId().getIdopcion(),
-								obj.getDescripcion(), obj.getPrioridad(), obj.getImagen(), obj.getOpcion() });
+								obj.getDescripcion(), op, obj.getImagen(), obj.getOpcion() });
 			}
 		} else {
 			txtCodigo.setText("");
@@ -356,6 +432,12 @@ public class FrmSysGrupo extends AbstractMaestro {
 		setSysGrupos(getSysGrupoDAO().findAll());
 	}
 
+	public void llenar_contenedores() {
+		ctnmodulo.refrescar();
+		ctntitulo.setModulo(null);
+		ctntitulo.refrescar();
+	}
+	
 	@Override
 	public void vista_edicion() {
 		if (getEstado().equals(NUEVO))
@@ -388,11 +470,15 @@ public class FrmSysGrupo extends AbstractMaestro {
 
 	@Override
 	public void eliminar() {
-		if (getSysGrupo() != null) {
-			// getAlmacenDAO().borrarPorSucursal(getSucursal());
-			getSysGrupoDAO().remove(getSysGrupo());
+		int seleccion = UtilMensajes.msj_error("ELIMINAR_REG");
+		if (seleccion == 0){	
+			if (getSysGrupo() != null) {
+				getSysOpcionDAO().borrarPorGrupo(getSysGrupo());
+				getSysGrupoDAO().remove(getSysGrupo());
+				
+				iniciar();
+			}
 		}
-
 	}
 
 	private DSGTableModel getDetalleTM() {
@@ -419,6 +505,14 @@ public class FrmSysGrupo extends AbstractMaestro {
 		this.sysGrupo = sysGrupo;
 	}
 
+	public SysGrupoPK getSysGrupoPK() {
+		return sysGrupoPK;
+	}
+
+	public void setSysGrupoPK(SysGrupoPK sysGrupoPK) {
+		this.sysGrupoPK = sysGrupoPK;
+	}
+
 	public SysModulo getSysModulo() {
 		return sysModulo;
 	}
@@ -439,6 +533,14 @@ public class FrmSysGrupo extends AbstractMaestro {
 		return sysOpciones;
 	}
 
+	public SysGrupo getSysGrupo2() {
+		return sysGrupo2;
+	}
+
+	public void setSysGrupo2(SysGrupo sysGrupo2) {
+		this.sysGrupo2 = sysGrupo2;
+	}
+
 	public void setSysOpciones(List<SysOpcion> sysOpciones) {
 		this.sysOpciones = sysOpciones;
 	}
@@ -457,5 +559,13 @@ public class FrmSysGrupo extends AbstractMaestro {
 
 	public void setSysOpcionDAO(SysOpcionDAO sysOpcionDAO) {
 		this.sysOpcionDAO = sysOpcionDAO;
+	}
+
+	public SysTitulo getSysTitulo() {
+		return sysTitulo;
+	}
+
+	public void setSysTitulo(SysTitulo sysTitulo) {
+		this.sysTitulo = sysTitulo;
 	}
 }
